Cherry pick sm4 encryption (#164)
* add sm4 encryption


* Update etc/config-template.toml


* *: update jemalloc to 5.3.0 (#12661)

close tikv/tikv#12660



* Adjust dependencies for SM4 encryption


* adapt code to new dependency

* fix ffi

* format

* fix ffi ci

* fix test

+++ src/coprocessor/tracker.rs
@@ -197,9 +197,9 @@ impl Tracker {
         let mut exec_details = kvrpcpb::ExecDetails::default();
 
         let mut td = kvrpcpb::TimeDetail::default();
-        td.set_process_wall_time_ms(time::duration_to_ms(measure) as i64);
-        td.set_wait_wall_time_ms(time::duration_to_ms(self.wait_time) as i64);
-        td.set_kv_read_wall_time_ms(self.scan_process_time_ms as i64);
+        td.set_process_wall_time_ms(time::duration_to_ms(measure));
+        td.set_wait_wall_time_ms(time::duration_to_ms(self.wait_time));
+        td.set_kv_read_wall_time_ms(self.scan_process_time_ms);
         exec_details.set_time_detail(td.clone());
 
         let detail = self.total_storage_stats.scan_detail();
